add logic to auto-play next tune when the user swipes to next song while isPlaying is true.
mediaId parameter is added to the AudioItem model. This will be used also to open the detail screen directly from the notification.
The idea is to format the media id like songbook||entry. Then we can open to the song entry directly. TBD
We use the mediaId also to determine when the current page has changed.

add DummyPlaybackController for android composable preview