#87: refactor PreferencesRepository to make it more generic and reusable
We should not change the repository for every new preference we want to add. Think of Open/Closed principle

Each new preference or settings simply instantiates and new key. A new usecase could be create if it needs some custom transformation. Otherwise, the PreferenceKey and the current repository should be enough (hopefully)